JX NEXT: anh test e2e sau M12 B3c + Lua KillPlayer 0x08117BC0 + PKRate.ini rate + goi 1111/1112/1113 bi chan hai lop (gateway session.go + zone KGameServer.cpp) - sua
- Lua KillPlayer 0x08117BC0 doc tung dong: 20 o KMagicAttrib (seriesdamage_p 100, attackrating_v 50000, ignoredefense_p 1, o vat ly 200 000 000) -> KNpc::ReceiveDamage(chinh minh, series 0, khong can chien, khong kiem AR, nDoHurt 1, quan he 0x1f); zone l_KillPlayer; KillNpc/KillNpcWithIdx -> 0x08117CE0 chua port
- [0x8BADF50] = g_NpcSet 0x8BACAC0 + 0x1490 = PKRate.ini [PK] rate (KNpcSet::Init 0x080A0810, mac dinh 20, tep 20); CalcDamage nhan 32-bit tran voi KillPlayer -> zone nhan 64-bit, pk_damage_percent 100 -> 20; chin khoa PK con lai ghi 16.4 cho B3c-4
- gateway session.go va zone KGameServer.cpp deu chan C2G_ADD_SKILL_POINT/C2G_CAST_SKILL/C2G_REVIVE (message not allowed in state / unhandled client message) -> them
- client --auto: buoc chet -> hoi sinh (AUTO_DEATH, auto_death.png / auto_revive.png); buoc vut do chon do trong tui khong phai vat pham nhiem vu (truoc: RESULT_BAD_REQUEST)
- test_KNpcCommand +1 ca (KillPlayer qua ?gm ds, 14 kiem), ctest 210/210, go test gateway, Godot 262/262; e2e AUTO_DROP dropped=true AUTO_DEATH dead=true revived=true
- LINUX-SERVER.md 16.4 + hang 0x08089C90, OLD-TO-NEW, HANDOVER 0.3/0.4/0.5/4b phan 11

+++ next/server/zone/src/KNpc.cpp
@@ -687,8 +687,10 @@ int KSubWorld::calc_damage(KNpc& t, KNpc& a, int min, int max, int type, bool me
         }
         calc_damage(a, t, ret, ret, damage_return, melee, nullptr, nullptr, 0, is_return);
     }
-    // 0x08089E31: between players (and partners) the PK rate [0x8BADF50]
-    if (player_or_partner(t) && player_or_partner(a)) dmg = dmg * cfg_.pk_damage_percent / 100;
+    // 0x08089E31 / 0x0808A350: between players (and partners) the PK rate NpcSet+0x1490 ([0x8BADF50] =
+    // PKRate.ini [PK] rate, 20).  The binary multiplies in 32 bits, so the 200 000 000 of KillPlayer wraps
+    // negative and heals; the zone keeps the product whole (the only blow that large)
+    if (player_or_partner(t) && player_or_partner(a)) dmg = static_cast<int>(static_cast<std::int64_t>(dmg) * cfg_.pk_damage_percent / 100);
     // (0x08089E44: a player's damage counter +0x86ac / +0x86b0 - B3)
     t.last_damage_id = a.id;   // +0x1598
     if (type == damage_poison && c.poison_dec_mana_percent != 0) {   // 0x0808A0C8: Posion2Mana
